add new report button to the thank you page
